+++ src/components/calcite-dropdown/calcite-dropdown.e2e.ts
@@ -798,14 +798,9 @@ describe("calcite-dropdown", () => {
         <calcite-dropdown>
           <calcite-action slot="dropdown-trigger">Open dropdown</calcite-action>
           <calcite-dropdown-group selection-mode="single">
-            <calcite-dropdown-item id="item-1">
-              Dropdown Item Content
-            </calcite-dropdown-item>
-            <calcite-dropdown-item id="item-2" active>
-              Dropdown Item Content
-            </calcite-dropdown-item>
+            <calcite-dropdown-item id="item-1"> Dropdown Item Content</calcite-dropdown-item>
+            <calcite-dropdown-item id="item-2" active> Dropdown Item Content</calcite-dropdown-item>
           </calcite-dropdown-group>
-          </calcite-dropdown>
         </calcite-dropdown>
       `);
       const element = await page.find("calcite-dropdown");
@@ -1108,5 +1103,36 @@ describe("calcite-dropdown", () => {
         return scrollHeight > clientHeight || scrollWidth > clientWidth;
       })
     ).toBe(false);
+  });
+
+  it("dropdown wrapper shouldn't have height when filter results empty and combined with a PickList in Panel  #3048", async () => {
+    const page = await newE2EPage({
+      html: html` <calcite-panel heading="Issue #3048">
+        <calcite-pick-list filter-enabled>
+          <calcite-dropdown slot="menu-actions" placement="bottom-trailing" type="click">
+            <calcite-action slot="dropdown-trigger" title="Sort" icon="sort-descending"> </calcite-action>
+            <calcite-dropdown-group selection-mode="single">
+              <calcite-dropdown-item>Display name</calcite-dropdown-item>
+              <calcite-dropdown-item>Type</calcite-dropdown-item>
+            </calcite-dropdown-group>
+          </calcite-dropdown>
+          <calcite-pick-list-item label="calcite" description="calcite!"> </calcite-pick-list-item>
+          <calcite-pick-list-item label="calcite" description="calcite"> </calcite-pick-list-item>
+        </calcite-pick-list>
+      </calcite-panel>`
+    });
+    await page.waitForChanges();
+
+    const dropdownContentHeight = await (
+      await page.find("calcite-dropdown >>> .calcite-dropdown-wrapper")
+    ).getComputedStyle();
+
+    await page.evaluate(() => {
+      const filter = document.querySelector(`calcite-pick-list`).shadowRoot.querySelector("calcite-filter");
+      const filterInput = filter.shadowRoot.querySelector("calcite-input");
+      filterInput.value = "nums";
+    });
+
+    expect(dropdownContentHeight.height).toBe("0px");
   });
 });